Split 'auth' model/config into 'token' and 'user'.
