+++ rest-server/src/controllers/job.js
@@ -15,11 +15,9 @@
 // DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-
 // module dependencies
 const Job = require('../models/job');
 const logger = require('../config/logger');
-
 
 /**
  * Load job and append to req.
