+++ rest-server/src/models/token.js
@@ -15,23 +15,17 @@
 // DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+const userModel = require("./user");
 
-// module dependencies
-const express = require('express');
-const tokenConfig = require('../config/token');
-const userConfig = require('../config/user');
-const userController = require('../controllers/user');
-const param = require('../middlewares/parameter');
+const check = (username, password, callback) => {
+  if (!userModel.db.has(username).value()) {
+    callback(null, false, false);
+  } else {
+    const user = userModel.db.get(username).value();
+    userModel.encrypt(username, password, (err, derivedKey) => {
+      callback(err, derivedKey === user.passwd, user.admin);
+    });
+  }
+};
 
-
-const router = express.Router();
-
-router.route('/')
-    /** PUT /api/v1/user - Create or update a user */
-    .put(tokenConfig.check, param.validate(userConfig.schema), userController.update)
-
-    /** DELETE /api/v1/user - Remove a user */
-    .delete(tokenConfig.check, userController.remove);
-
-// module exports
-module.exports = router;
+module.exports = { check };
